pyproject.toml
[tool.poetry]
name = "my-app"
version = "0.1.0"
description = "Project template for Django app, using Poetry."
license = "MIT"
authors = ["YunoJuno <code@yunojuno.com>"]
maintainers = ["YunoJuno <code@yunojuno.com>"]
readme = "README.md"
homepage = "https://github.com/yunojuno/poetry-template"
repository = "https://github.com/yunojuno/poetry-template"
documentation = "https://github.com/yunojuno/poetry-template"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
packages = [{ include = "my_app" }]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2 || ^4.0 || ^5.0"
django-ninja = {version = "1.0b1", allow-prereleases = true}
django-microsoft-auth = "^3.0.1"
django-extensions = "^3.2.3"

[tool.poetry.dev-dependencies]
black = "*"
coverage = "*"
freezegun = "*"
mypy = "*"
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
ruff = "*"
tox = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
